Création des pages de profil (on peut mettre à jour ses infos depuis cette dite page) et mis à disposition de liens permettant d'y accéder. La page listant les signalements est maitenant liée à celle listant les utilisateurs au travers des ancres. Il est maintenant possible de supprimer un billet depuis la page des signalements. Le format des dates a été revu et est plus cohérent.Utilisation de boostrap pour améliorer les liens liés à la pagination. Un dossier contenant les images est créé lors de l'inscription d'un utilisateur. Il est bien détruit ainsi que tous les fichiers qu'il contient lors de la suppression de ce dernier. Inversion des pages connexion et formulaireconnexion via l'urlrewritting. Verrouillage d'accès des pages ET des actions non autorisées. Redirection vers une page dédiée lors d'une erreur 404.

+++ index.php
@@ -5,9 +5,9 @@
         include('View/SignUp.php');
     }elseif(filter_input(INPUT_GET, 'action') == 'signInView') {
         include('View/SignIn.php');
-    }elseif(filter_input(INPUT_POST, 'email')) {
+    }elseif(filter_input(INPUT_GET, 'action') == 'signUp') {
         signUp();
-    }elseif(filter_input(INPUT_POST, 'pseudo')) {
+    }elseif(filter_input(INPUT_GET, 'action') == 'signIn') {
         signIn();
     }elseif(filter_input(INPUT_GET, 'action') == 'ticketsList') {
         ticketsList();
@@ -22,25 +22,46 @@
     }elseif(filter_input(INPUT_GET, 'action') == 'report') {
         createReport();
     }elseif(filter_input(INPUT_GET, 'action') == 'addTicketView') {
-        include('View/AddTicket.php');
+        if(authorized()) {
+            include('View/AddTicket.php');
+        }
     }elseif(filter_input(INPUT_GET, 'action') == 'addTicket') {
         addTicket();
     }elseif(filter_input(INPUT_GET, 'action') == 'updateTicket') {
         updateTicket();
     }elseif(filter_input(INPUT_GET, 'action') == 'deleteTicket') {
-        deleteTicket();
+        if(authorized()) {
+            deleteTicket();
+        }
     }elseif(filter_input(INPUT_GET, 'action') == 'signOut') {
         signOut();
     }elseif(filter_input(INPUT_GET, 'action') == 'usersList') {
-        usersList();
+        if(authorized()) {
+            usersList();
+        }
     }elseif(filter_input(INPUT_GET, 'action') == 'reportsList') {
-        reportsList();
+        if(authorized()) {
+            reportsList();
+        }
     }elseif(filter_input(INPUT_GET, 'action') == 'updateStatus') {
-        updateStatus();
+        if(authorized()) {
+            updateStatus();
+        }
     }elseif(filter_input(INPUT_GET, 'action') == 'deleteUser') {
-        deleteUser();
+        if(authorized()) {
+            deleteUser();
+        }
     }elseif(filter_input(INPUT_GET, 'action') == 'profilePage') {
         profilePage();
+    }elseif(filter_input(INPUT_GET, 'action') == 'updateEmail') {
+        updateEmail();
+    }elseif(filter_input(INPUT_GET, 'action') == 'updatePassword') {
+        updatePassword();
+    }elseif(filter_input(INPUT_GET, 'action') == 'updateImage') {
+        updateImage();
+    }
+    elseif(filter_input(INPUT_GET, 'action') == '404') {
+        error404();
     }
     else {
         ticketsList();
